backend/requirements.txt
aiofiles==24.1.0
bcrypt==4.2.0
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
PyJWT==2.10.0
python-dotenv==1.0.0
python-multipart==0.0.17
pydantic==2.4.2
python-slugify==8.0.1